Move WarningsSymbol to own component
